refactor: main.js, organize js files
